Add Factor class and some child implementations
The idea is to create classes that follow the order of the operations.
In the top level we have a polynomial with multiple terms. These terms
should be only combined in the last step. Each term consists of factors,
which are combined before the the terms. Powers and radicals should be
calculated in the first step.